Inline blob storage in news functions, remove newsStore lib
Both coindesk-news and polygon-news now use @netlify/blobs directly
instead of going through the newsStore.cjs wrapper. This matches the
pattern used by vend-blobs and removes an unnecessary abstraction layer.
